+++ src/main/java/com/prenotapp/_model/Appointment.java
@@ -1,7 +1,5 @@
 package com.prenotapp._model;
 
-import java.time.LocalDateTime;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.ForeignKey;
@@ -11,6 +9,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -26,14 +25,24 @@ public class Appointment {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Long id;
 
   @ManyToOne
-  @JoinColumn(name = "persona_id", referencedColumnName = "id", nullable = false, foreignKey = @ForeignKey(name = "fk_appointments_persona_id"))
+  @JoinColumn(
+    name = "persona_id",
+    referencedColumnName = "id",
+    nullable = false,
+    foreignKey = @ForeignKey(name = "fk_appointments_persona_id")
+  )
   private Persona persona;
 
   @ManyToOne
-  @JoinColumn(name = "shop_id", referencedColumnName = "id", nullable = false, foreignKey = @ForeignKey(name = "fk_appointments_shop_id"))
+  @JoinColumn(
+    name = "shop_id",
+    referencedColumnName = "id",
+    nullable = false,
+    foreignKey = @ForeignKey(name = "fk_appointments_shop_id")
+  )
   private Shop shop;
 
   @Column(name = "date", nullable = false)
