+++ src/main/java/com/prenotapp/_repo/IPersonaRepo.java
@@ -2,4 +2,4 @@ package com.prenotapp._repo;
 
 import com.prenotapp._model.Persona;
 
-public interface IPersonaRepo extends IGenericRepo<Persona, Integer> {}
+public interface IPersonaRepo extends IGenericRepo<Persona, Long> {}
