+++ src/main/java/com/prenotapp/_service/IAppointmentService.java
@@ -1,14 +1,13 @@
 package com.prenotapp._service;
 
-import java.util.List;
-
 import com.prenotapp._dto.AppointmentDetailsDTO;
 import com.prenotapp._model.Appointment;
+import java.util.List;
 
-public interface IAppointmentService extends ICRUD<Appointment, Integer> {
+public interface IAppointmentService extends ICRUD<Appointment, Long> {
   public List<AppointmentDetailsDTO> listAppointments() throws Exception;
 
-  public AppointmentDetailsDTO findAppointmentById(Integer idAppointment)
+  public AppointmentDetailsDTO findAppointmentById(Long idAppointment)
     throws Exception;
 
   public boolean appointmentAlreadyExist(Appointment appointment)
