+++ src/main/java/com/prenotapp/_service/IShopCategoryService.java
@@ -2,4 +2,4 @@ package com.prenotapp._service;
 
 import com.prenotapp._model.ShopCategory;
 
-public interface IShopCategoryService extends ICRUD<ShopCategory, Integer> {}
+public interface IShopCategoryService extends ICRUD<ShopCategory, Long> {}
